openwrt compatibale

+++ build.sh
@@ -33,6 +33,7 @@ if [ -z $DEFAULT_SOURCE ]; then
 fi
 CPU_COUNT="$(cat /proc/cpuinfo | grep processor | wc -l)"
 REPO_NAME="$(cut -d \: -f 1 <<< ${DEFAULT_SOURCE} | cut -d \/ -f 2)"
+REPO_USER="$(cut -d \: -f 1 <<< ${DEFAULT_SOURCE} | cut -d \/ -f 1)"
 REPO_URL="https://github.com/$(cut -d \: -f 1 <<< ${DEFAULT_SOURCE})"
 REPO_BRANCH=$(cut -d \: -f 2 <<< ${DEFAULT_SOURCE})
 OPENWRT_BASE="${BUILD_WORKSPACE}/${REPO_NAME}"
@@ -81,12 +82,16 @@ unset p
 
 cd ${BUILD_WORKSPACE}
 if [ ! -d ${BUILD_WORKSPACE}/${REPO_NAME} ]; then
-    git clone -b ${REPO_BRANCH} --single-branch --depth 1 ${REPO_URL}.git
+    git clone -b ${REPO_BRANCH} --single-branch --depth 1 ${REPO_URL}.git ${REPO_NAME}
 fi
 
-cd ${BUILD_WORKSPACE}
-if [ $KERNEL_CONFIG == "x86_64" ]; then
-    git clone -b master https://github.com/openwrt/openwrt.git openwrt
+if [ "$REPO_USER" == "openwrt" ] && [ "$REPO_NAME" == "openwrt" ]; then
+    git clone -b master --single-branch --depth 1 https://github.com/coolsnowwolf/lede.git lede
+    cp -aRp lede/lean ${OPENWRT_BASE}/package/
+fi
+
+if [ $KERNEL_CONFIG == "x86_64" ] && [ "$REPO_USER" == "coolsnowwolf" ] && [ "$REPO_NAME" == "lede" ]; then
+    git clone -b master --single-branch --depth 1 https://github.com/openwrt/openwrt.git openwrt
     rm -rf ${OPENWRT_BASE}/package/kernel/mac80211
     cp -aRp openwrt/package/kernel/mac80211 ${OPENWRT_BASE}/package/kernel/    
 fi
